修复 solve 函数中的变量类型，将 int n 改为 i64 n，以支持更大的输入值

+++ program_2025/program_11/program_20/P_4718.cpp
@@ -106,7 +106,7 @@ void factorize(i64 n, vector<i64>& factors) {
 
 void solve()
 {
-    int n;
+    i64 n;
     cin >> n;
     if (is_prime(n)) {
         cout << "Prime\n";
